requirements.txt
Pillow>=9.0.0
tkinter-dnd2>=0.3.0